Disambiguate destructuring
There were a few cases that were too ambiguous for the parser. It is at least
temporarily solved by making the type destructures higher priority. The
following example was problematic before.

    fn: Int -> List a -> a b

The parser was interpreting that as a function from Int to a List of a->a
fucntions with b as a second parameter to List.