Disambiguate the grammar
There were serious performance issues with the first approach. These
changes expand the grammar to disambiguate it so that the parser
performs better by having less possibilities to evaluate. In addition,
the start rule does mot model itself after the structure of an elm file.
In stead, it is a series of definitions in no particular order.